Step 2 of stl refactorization
Done with the restructuring of the stl examples. Foo test class is now
in a file of its own. Same goes for the performace counter timer impl
used for benchmarks in windows. mapUsages incl multimap stuff is in a
file of its own and the same goes for vector example and benchmark
examples.
Fixed the benchmark code for removing 1000 elements from pos 1000 i a
list to not itterate unnessesary 1000*1000 steps.